requirements.txt
django-cors-headers==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
requests==2.26.0
python-dotenv==0.19.0
